ia64: tentative fix for multiple pending signals handling
* recheck stuff after do_notify_resume()
* use new TS_... flag to prevent multiple restarts
* clear that flag once we are on our way out to userland

It seems to work, but the assembler part is certainly an atrocity - as
it is it's a massive pipeline stall just to clear the damn bit in
current_thread_info()->status.  Needs a look from somebody who knows
how insn scheduling works on ia64; what I'd done there is as blunt
as it gets and it certainly can be handled better.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>