UBI: provide a write-pointer API
Writing on LEB is currently pretty easy since it only implies writing
sequentially on all of its write units, but the apparition of MLC/TLC
NANDs complexify the logic because of the paired pages concept.

The write pointer API is here to help UBI users with the constraints
brought by paired pages.
It provides helpers to secure written data, find the next contiguous memory
region, and move the write pointer.
It also keeps track of skipped pages so that the user can avoid writing on
it.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>