of: Add new device status value "okay-manually"
Introduce a new device tree status value, "okay-manually".

Similar to "okay" or "ok", this value indicates that the device
is available and functional. The key difference is that devices
marked as "okay-manually" are not automatically probed by the OS.
Instead, they must be explicitly bound to a driver via sysfs.

This is useful in setups where multiple drivers may support the
same device. Automatically binding the "wrong" driver first can be
problematic, as unbinding is often cumbersome and not always safe.
Some drivers leave devices in an undefined state after unbinding
which complicates repeated binding attempts.

In addition, avoiding unnecessary probes improves boot performance,
especially for complex devices that require firmware loading.

Signed-off-by: Richard Weinberger <richard@nod.at>