drbd: add DAX/PMEM support
In case the meta-data is on persistent memory (PMEM aka NVDIMM) and we
have direct access (DAX) to it, it does not make sense to access it only
by writing/reading blocks from/to it.
We can access it byte-granular, and access can happen from IRQ contexts,
and is nearly as fast as access to main memory.

Signed-off-by: Christoph Böhmwalder <christoph.boehmwalder@linbit.com>